fix(git-cli): ensure OS-agnostic file parent directory handling
This commit addresses an issue found within the amplication-git-utils package, specifically in the git-cli.ts file. 

Previously, the handling of the file parent directory (`fileParentDir`) was not OS-agnostic, causing issues particularly when running the software on Windows machines. This was manifesting as an ENOENT error, indicating that a required directory could not be created or found.

The changes in this commit include utilizing the `dirname` function from Node.js built-in 'path' module for extracting the parent directory from a file path, which provides a more consistent OS-agnostic solution, instead of spliting string with "/" .

This is a resolution for the GitHub issue #6149.