Update README.md
Added the target="_blank" attribute to the anchor tags (<a>) for the Discord, Twitter, and YouTube icons to open the links in new tabs. This change ensures that when users click on these icons, the linked pages will open in separate tabs, providing a better user experience and preventing users from navigating away from the current page.